refactor: Rename github_helper to github_datasets for consistency
